# api/openapi.yaml
openapi: 3.0.3
info:
  title: Travel Companion API
  version: 1.0.0
  description: |
    Travel Companion — connect travelers who need language assistance with
    those who can provide it on the same flight.
  contact:
    name: Travel Companion
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PostType:
      type: string
      enum: [SEEKING_ASSISTANCE, OFFERING_ASSISTANCE]
    PostSegmentDto:
      type: object
      required: [segment_order, flight_number, airline, origin_iata, destination_iata, departure_time, arrival_time]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        segment_order:
          type: integer
          minimum: 1
        flight_number:
          type: string
        airline:
          type: string
        origin_iata:
          type: string
        destination_iata:
          type: string
        departure_time:
          type: string
          format: date-time
        arrival_time:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      required: [post_type, origin_iata, final_destination_iata, travel_date, segments]
      properties:
        post_type:
          $ref: '#/components/schemas/PostType'
        poster_is_traveller:
          type: boolean
        traveller_relationship:
          type: string
        origin_iata:
          type: string
        final_destination_iata:
          type: string
        travel_date:
          type: string
          format: date
        languages_needed:
          type: array
          items:
            type: string
        languages_spoken:
          type: array
          items:
            type: string
        notes:
          type: string
          maxLength: 280
        segments:
          type: array
          items:
            $ref: '#/components/schemas/PostSegmentDto'
    PostResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        poster_id:
          type: string
        post_type:
          $ref: '#/components/schemas/PostType'
        poster_is_traveller:
          type: boolean
        traveller_relationship:
          type: string
        origin_iata:
          type: string
        final_destination_iata:
          type: string
        travel_date:
          type: string
          format: date
        languages_needed:
          type: array
          items:
            type: string
        languages_spoken:
          type: array
          items:
            type: string
        notes:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        segments:
          type: array
          items:
            $ref: '#/components/schemas/PostSegmentDto'
...
    ConnectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        initiator_id:
          type: string
        post_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted, declined, blocked]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateConnectionRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined, blocked]
    TravellerDetailsRequest:
      type: object
      required: [traveller_name]
      properties:
        traveller_name:
          type: string
        notes:
          type: string
    TravellerDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shared_by_user_id:
          type: string
        traveller_name:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
security:
  - BearerAuth: []
paths:
...
  /posts:
    post:
      tags: [Posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /posts/search:
    get:
      tags: [Posts]
      summary: Search posts
      parameters:
        - name: origin_iata
          in: query
          schema:
            type: string
        - name: destination_iata
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: languages
          in: query
          schema:
            type: array
            items:
              type: string
        - name: post_type
          in: query
          schema:
            $ref: '#/components/schemas/PostType'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
  /posts/{id}:
    get:
      tags: [Posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Not found
    patch:
      tags: [Posts]
      summary: Edit/deactivate own post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          description: Updated
    delete:
      tags: [Posts]
      summary: Delete own post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /posts/{id}/matches:
    get:
      tags: [Posts]
      summary: Find posts with overlapping flight segments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
  /posts/{id}/connections:
    post:
      tags: [Connections]
      summary: Initiate a connection request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Connection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
        '409':
          description: Already connected
  /connections:
    get:
      tags: [Connections]
      summary: List own connections
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, declined, blocked]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionResponse'
  /connections/{id}:
    patch:
      tags: [Connections]
      summary: Accept / decline / block a connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConnectionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionResponse'
  /connections/{id}/traveller-details:
    post:
      tags: [Connections]
      summary: Share traveller details after acceptance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TravellerDetailsRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravellerDetailsResponse'
...